fixed deleting default image from disk

+++ app/Http/Controllers/ProfileController.php
@@ -35,7 +35,7 @@ class ProfileController extends Controller
 
         // If we have new image, delete old image and save the new one
         if ($request->hasFile('photo')){
-            if ($user->photo){
+            if ($user->photo && $user->photo != 'user_logo/default-image.png' ){
                 Storage::disk('public')->delete($user->photo);
             }
             $path = Storage::disk('public')->put('user_logo', $request->photo);
